requirements.txt
numpy
pyaudio
aubio
paho-mqtt
rich
textual
pygame